# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/subscription": {
            "get": {
                "description": "Returns a list of subscription objects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Read subscription list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.LoadSubResponce"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorBadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorInternalError"
                        }
                    }
                }
            },
            "post": {
                "description": "Returns a new subscription object.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Create new subscription",
                "parameters": [
                    {
                        "description": "Subscription create data",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorBadRequest"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorInternalError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Returns an ID updated subscription.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Update subscription by ID",
                "parameters": [
                    {
                        "description": "Subscription update data",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSubRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSubResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorBadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorInternalError"
                        }
                    }
                }
            }
        },
        "/subscription/cost": {
            "post": {
                "description": "Returns a cost of subscriptions by user ID, date and service name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Cost subscription",
                "parameters": [
                    {
                        "description": "Subscription update data",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CostResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorBadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorInternalError"
                        }
                    }
                }
            }
        },
        "/subscription/{id}": {
            "get": {
                "description": "Returns a subscription object.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Read subscription by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Subscription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoadSubResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorBadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorInternalError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Returns an ID deleted subscription.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Delete subscription by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Subscription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteSubResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorBadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorInternalError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CostRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "02-2025"
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "01-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "dceb1963-e152-47ff-a562-81a360627309"
                }
            }
        },
        "dto.CostResponce": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "integer",
                    "example": 900
                },
                "months_count": {
                    "type": "integer",
                    "example": 3
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "user_id": {
                    "type": "string",
                    "example": "dceb1963-e152-47ff-a562-81a360627309"
                }
            }
        },
        "dto.CreateSubRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "02-2025"
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "01-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "dceb1963-e152-47ff-a562-81a360627309"
                }
            }
        },
        "dto.CreateSubResponce": {
            "type": "object",
            "properties": {
                "subscription_id": {
                    "type": "integer",
                    "example": 123
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.DeleteSubResponce": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.LoadSubResponce": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "02-2025"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "01-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "dceb1963-e152-47ff-a562-81a360627309"
                }
            }
        },
        "dto.UpdateSubRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "07-2025"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "type": "integer",
                    "example": 399
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Minus"
                },
                "start_date": {
                    "type": "string",
                    "example": "05-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "dceb1963-e152-47ff-a562-81a360627309"
                }
            }
        },
        "dto.UpdateSubResponce": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handler.ErrorBadRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error text"
                },
                "status": {
                    "type": "string",
                    "example": "bad request"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handler.ErrorInternalError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error text"
                },
                "status": {
                    "type": "string",
                    "example": "internal error"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handler.ErrorNotFound": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error text"
                },
                "status": {
                    "type": "string",
                    "example": "entity not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        }
    }
}